pyproject.toml
[project]
name = "elabftw-mcp"
version = "0.1.0"
description = "MCP Server for elabFTW - Electronic Lab Notebook"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.25.0",
    "anyio>=4.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
only-include = ["elabftw_mcp_server.py"]

[tool.hatch.build.targets.sdist]
only-include = ["elabftw_mcp_server.py", "pyproject.toml", "README.md"]
